made user types public
