+++ .github/workflows/linters.yml
@@ -4,7 +4,7 @@ on: [push, pull_request]
 
 jobs:
   clippy:
-    name: Clippy
+    name: clippy
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
