+++ src/user/mod.rs
@@ -1,5 +1,3 @@
 mod api;
-pub use api::{
-	Invitation, ResUserPasswordChange, User, UserApi, UserAuth, UserAuthLoginOptions,
-	UserCreateConfig, UserInfo,
-};
+pub use api::UserApi;
+pub mod types;
